+++ cogs/core.py
@@ -2,6 +2,7 @@ from discord.ext import commands
 from discord.utils import get
 import discord, traceback, json, datetime
 from ext.imports_share import chk_sudo
+import typing
 BOTSETFILE = "ext/bot_settings.json"
 SETFILE = "data/settings.json"
 settings = json.load(open(SETFILE, 'r'))
@@ -24,6 +25,7 @@ class Core(commands.Cog):
     - reload
     - load
     """
+
     def __init__(self, bot):
         self.bot = bot
 
@@ -149,7 +151,97 @@ class Core(commands.Cog):
             await ctx.send(embed=embed)
             return
 
+    @info.command(name='user', help='info about a user (can be outside of this server!)')
+    async def info_user(self, ctx: commands.Context, user: discord.User = None):
+        user = user or ctx.author
+        other_desc = ""
+        if user.discriminator:
+            other_desc += f":warning: username has conflict: {user.discriminator}\n"
+        embed = discord.Embed(title=f'user {user.display_name}', description=f'```{user.id}{" | BOT" if user.bot else ""}```')
+        embed.add_field(name='Mention', value=f"{user.mention} / `{user.mention}`")
+
+        embed.set_author(name=user, icon_url=user.avatar_url)
+        embed.set_footer(text='Created')
+        embed.timestamp = user.created_at
+        await ctx.send(embed=embed)
+        return 0
+
+    @info.command(name='member', help='info about a member')
+    @commands.guild_only()
+    async def info_member(self, ctx, member: discord.Member = None):
+        member = member or ctx.author
+        other_desc = ""
+        if member.discriminator:
+            other_desc += f":warning: username has conflict: {member.discriminator}\n"
+        embed = discord.Embed(title=f'Member {member.display_name}', description=f'```{member.status} | {member.id}{" | BOT" if member.bot else ""}```')
+        embed.add_field(name='Mention', value=f"{member.mention} / `{member.mention}`")
+        if member.is_on_mobile():
+            embed.add_field(name='Device', value='Mobile')
+            if member.mobile_status == discord.Status.online:
+                embed.add_field(name='Mobile Status', value='online')
+            elif member.mobile_status == discord.Status.idle:
+                embed.add_field(name='Mobile Status', value='idle')
+            elif member.mobile_status == discord.Status.dnd:
+                embed.add_field(name='Mobile Status', value='do not disturb')
+            else:
+                embed.add_field(name='Mobile Status', value='offline / invisible')
+        else:
+            embed.add_field(name='Device', value='Desktop / Web App')
+
+        if member.desktop_status == discord.Status.online:
+            embed.add_field(name='Desktop status', value='online')
+        elif member.desktop_status == discord.Status.idle:
+            embed.add_field(name='Desktop Status', value='idle')
+        elif member.desktop_status == discord.Status.dnd:
+            embed.add_field(name='Desktop Status', value='do not disturb')
+        else:
+            embed.add_field(name='Desktop Status', value='offline / invisible')
+
+        if member.web_status == discord.Status.online:
+            embed.add_field(name='Web App status', value='online')
+        elif member.web_status == discord.Status.idle:
+            embed.add_field(name='Web App Status', value='idle')
+        elif member.web_status == discord.Status.dnd:
+            embed.add_field(name='Web App Status', value='do not disturb')
+        else:
+            embed.add_field(name='Web App Status', value='offline / invisible')
+
+        embed.set_author(name=member, icon_url=member.avatar_url)
+        embed.add_field(name='Nickname', value=member.nick or f"<{member.mention} have no nickname>")
+        embed.add_field(name='Roles', value=', '.join([r.mention for r in member.roles[1:]]))
+        embed.set_footer(text='Joined', icon_url=ctx.guild.icon_url)
+        embed.timestamp = member.joined_at
+        await ctx.send(embed=embed)
+        return 0
+
+    @info.command(name='channel', help='info about a channel')
+    @commands.guild_only()
+    async def info_channel(self, ctx, channel: discord.TextChannel = None):
+        channel = channel or ctx.channel
+        other_desc = ""
+        if channel.is_nsfw():
+            other_desc += ":warning: nsfw channel\n"
+        if channel.is_news():
+            other_desc += ":newspaper: news channel\n"
+        if channel.category is not None:
+            if channel.permissions_synced:
+                other_desc += ":arrows_counterclockwise: :white_check_mark: Permission synced"
+            else:
+                other_desc += ":arrows_counterclockwise: :negative_squared_cross_mark: Permission outdated"
+        embed = discord.Embed(title=f'Text Channel {channel.name}', description=f"```{channel.id}```{other_desc}")
+        embed.add_field(name="Category Position", value=channel.position or "<No position>")
+        embed.add_field(name='Category', value="<GLOBAL>" if channel.category is None else channel.category.name)
+        embed.add_field(name='Mention', value=f"{channel.mention} / `{channel.mention}`")
+        if any(await channel.invites()):
+            embed.add_field(name='Invites', value=", ".join(f"[{invite.id}]({invite.url})" for invite in channel.invites()))
+        embed.add_field(name='Pinned messages', value=len(await channel.pins()))
+        embed.set_footer(text="Channel created")
+        embed.timestamp = channel.created_at
+        await ctx.send(embed=embed)
+        return 0
+
     @info.command(name='server', help='info about the current server', aliases=['guild', 'srv', 'g'])
+    @commands.guild_only()
     async def info_server(self, ctx):
         settings = json.load(open(SETFILE, 'r'))
         embed = discord.Embed(title='Server info', description=ctx.guild.description or "<description not set>")