+++ cogs/utils.py
@@ -6,7 +6,6 @@ import random
 import botmc
 import traceback
 from datetime import datetime
-SANDBOX_TRACEBACK = 'samples/traceback.sndbx'
 
 
 class Utils(commands.Cog):
@@ -19,6 +18,8 @@ class Utils(commands.Cog):
     ## Commands  
     - vote / poll
     - mc
+    - invite
+    - avatar
     """
     def __init__(self, bot):
         self.bot = bot
@@ -183,56 +184,12 @@ class Utils(commands.Cog):
         await ctx.send(f.read())
 
     @commands.command(name='invite', help='get server invite link')
+    @commands.guild_only
     async def invite(self, ctx):
         await ctx.send((await ctx.guild.invites())[0].url)
 
-    @commands.command(name='sandbox', help='check if a command runs properly')
-    async def sandbox(self, ctx, *, commandName: str):
-        command = None
-        ret = None
-        startTime = None
-        errorCode = int()
-        try:
-            command = self.bot.get_command(commandName)
-            if command is None:
-                await ctx.send(f"msh: command not found: `{commandName}`")
-                return 3
-            startTime = datetime.now()
-            ret = await ctx.invoke(command)
-        except Exception as err:
-            timeElapsed = datetime.now() - startTime
-            e = discord.Embed(title='Task Failed Succefully', description=f":x: `{ctx.message.content.split()[0][:-7]}{commandName}`", color=0xff0000)
-            e.add_field(name='command name', value=command.qualified_name)
-            e.add_field(name='time used', value=timeElapsed)
-            try:
-                errorCode = str(int(ret))
-            except Exception:
-                try:
-                    errorCode = str(ret)
-                except Exception:
-                    errorCode = str("return value cannot be converted to string")
-            e.add_field(name='Return value', value=errorCode)
-            e.add_field(name='Error type', value=type(err).__name__)
-            e.add_field(name='Error message', value=str(err))
-            e.timestamp = datetime.utcnow()
-            with open(SANDBOX_TRACEBACK, 'w') as f:
-                f.write(traceback.format_exc())
-            await ctx.send(embed=e, file=discord.File(open(SANDBOX_TRACEBACK, 'r'), 'traceback.txt'))
-            return 0
-        timeElapsed = datetime.now() - startTime
-        e = discord.Embed(title='the command runs without any error!', description=f":white_check_mark: `{ctx.message.content.split()[0][:-7]}{commandName}`", color=0x00ff00)
-        e.add_field(name='command name', value=command.qualified_name)
-        e.add_field(name='time used', value=timeElapsed)
-        try:
-            errorCode = str(int(ret))
-        except Exception:
-            try:
-                errorCode = str(ret)
-            except Exception:
-                errorCode = str("return value cannot be converted to string")
-        e.add_field(name='Return value', value=errorCode)
-        e.timestamp = datetime.utcnow()
-        await ctx.send(embed=e)
+    @commands.command(name='avatar')
+    async def 
 
 
 def setup(bot):